refactor: ソートオプション定数をconstantsファイルに分離
SortSelectorとusePracticePageで重複していたSORT_OPTIONS・
DIFFICULTY_ORDER・SortOption型をconstants/sortOptions.tsに統合。

closes #697
